Move query builders to search module
These should live with the other search query
classes.

I've just noticed there aren't unit tests for
these classes.
